took the form off of the application views, and moved it to the list.
